requirements.txt
# # Dependencias del proyecto
Flask>=3.0,<4
mysql-connector-python==9.0.0
python-dotenv==1.0.1